+++ src/Models/Currency.php
@@ -44,4 +44,5 @@ class Currency
     public const CUR_RSD = 'RSD'; // Serbia
     public const CUR_UAH = 'UAH'; // Ukraine
     public const CUR_GEL = 'GEL'; // Georgia
+    public const CUR_MDL = 'MDL'; // Moldavia
 }